pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ismip6-helper"
version = "0.1.0"
description = "Helper utilities for loading ISMIP6 outputs"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Thomas Teisberg" },
]
keywords = ["ismip6", "ice-sheet", "climate", "xarray", "panel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "cftime>=1.6.5",
    "dask>=2025.11.0",
    "fsspec>=2025.9.0",
    "gcsfs>=2025.9.0",
    "h5netcdf>=1.7.3",
    "holoviews>=1.19.0",
    "hvplot>=0.10.0",
    "ipywidgets>=8.1.0",
    "jupyter>=1.1.0",
    "jupyter-bokeh>=4.0.5",
    "matplotlib>=3.10.7",
    "netcdf4>=1.7.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "panel>=1.5.4",
    "pyarrow>=22.0.0",
    "pyproj>=3.7.0",
    "pyyaml>=6.0.0",
    "scipy>=1.16.2",
    "xarray>=2025.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ismip6_helper", "ismip6_comparison_app"]
